moved getUniqRandomItemsFromArray func to separate file
